requirements.txt
python-telegram-bot
psycopg
psycopg_binary
